requirements.txt
pkg-resources==0.0.0
python-gnupg==0.4.5
